Fix how players pick a card from the deck.
Players should not be picking a random card, but rather, the next available card, - i.e., the top card - from the deck. This ensures the last player will always pick the initial lead - or briscola - card, which is put back at the end of the deck at the start of the game.